documenting and merging webcache and webcache_cookies
