+++ plaso/parsers/esedb_plugins/__init__.py
@@ -3,6 +3,5 @@
 
 from plaso.parsers.esedb_plugins import file_history
 from plaso.parsers.esedb_plugins import msie_webcache
-from plaso.parsers.esedb_plugins import msie_webcache_cookies
 from plaso.parsers.esedb_plugins import srum
 from plaso.parsers.esedb_plugins import user_access_logging
