+++ plaso/parsers/esedb_plugins/msie_webcache.py
@@ -13,6 +13,40 @@ from plaso.parsers import esedb
 from plaso.parsers.esedb_plugins import interface
 
 
+class MsieWebCacheCookieData(events.EventData):
+  """MSIE WebCache Container table event data.
+
+  Attributes:
+    container_identifier (int): container identifier.
+    cookie_hash (str): a similarity hash of the cookie contents
+    cookie_name (str): name of the cookie
+    cookie_value_raw (str): raw value of cookie in hex
+    cookie_value (str): value of the cookie encoded in ascii
+    entry_identifier (int): entry identifier.
+    expiration_time (dfdatetime.DateTimeValues): expiration date and time.
+    flags (int): an representation of cookie flags
+    modification_time (dfdatetime.DateTimeValues): modification date and time.
+    request_domain (str): Request domain for which the cookie was set.
+  """
+
+  DATA_TYPE = 'msie:cookie:entry'
+
+  def __init__(self):
+    """Initializes event data."""
+    super(MsieWebCacheCookieData, self).__init__(
+        data_type=self.DATA_TYPE)
+    self.container_identifier = None
+    self.cookie_hash = None
+    self.cookie_name = None
+    self.cookie_value_raw = None
+    self.cookie_value = None
+    self.entry_identifier = None
+    self.expiration_time = None
+    self.flags = None
+    self.modification_time = None
+    self.request_domain = None
+
+
 class MsieWebCacheContainersEventData(events.EventData):
   """MSIE WebCache Containers table event data.
 
@@ -278,6 +312,79 @@ class MsieWebCacheESEDBPlugin(interface.ESEDBPlugin):
 
         parser_mediator.ProduceEventData(event_data)
 
+  def _CookieHexToAscii(self, raw_cookie):
+    """Translates a cookie from a bytestring to a string
+
+    Args:
+      raw_cookie (bytes): the raw bytestring of a cookie field
+
+    Returns:
+      optional(str): the bytestring decoded and trimmed or None
+    """
+    if raw_cookie is not None:
+      raw_cookie = raw_cookie.rstrip(b'\x00')
+      return raw_cookie.decode('utf-8')
+    return None
+
+  def GetRawCookieValue(self, record_values, value_name):
+    """Retrieves the binary string as a hex string
+      
+    Args:
+      record_values: esedb dict structure for a table entry
+      value_name: the name of the value we are converting
+
+    Returns:
+      optional(str): the value of value_name in record_values decoded to hex
+    """
+    cookie_hash = record_values.get(value_name, None)
+    if cookie_hash is not None:
+      return f"0x{cookie_hash.hex()}"
+    return None
+
+  def _ParseCookieExTable(self, parser_mediator, table):
+    """Parses a CookieEntryEx_# table.
+
+    Args:
+      parser_mediator (ParserMediator): mediates interactions between parsers
+          and other components, such as storage and dfVFS.
+      table (pyesedb.table): table.
+      container_name (str): container name, which indicates the table type.
+    """
+    for record_index, esedb_record in enumerate(table.records):
+      if parser_mediator.abort:
+        break
+
+      try:
+        record_values = self._GetRecordValues(
+            parser_mediator, table.name, record_index, esedb_record)
+
+      except UnicodeDecodeError:
+        parser_mediator.ProduceExtractionWarning((
+            'Unable to retrieve record values from record: {0:d} '
+            'in table: {1:s}').format(record_index, table.name))
+        continue
+
+      cookie_name = self._CookieHexToAscii(record_values.get('Name', None))
+      cookie_value = self._CookieHexToAscii(record_values.get('Value', None))
+
+      cookie_hash = self.GetRawCookieValue(record_values, 'CookieHash')
+      cookie_value_raw = self.GetRawCookieValue(record_values, 'Value')
+
+      event_data = MsieWebCacheCookieData()
+      event_data.container_identifier = record_values.get('ContainerId', None)
+      event_data.cookie_hash = cookie_hash
+      event_data.cookie_name = cookie_name
+      event_data.cookie_value_raw = cookie_value_raw
+      event_data.cookie_value = cookie_value
+      event_data.entry_identifier = record_values.get('EntryId', None)
+      event_data.flags = record_values.get('Flags', None)
+      event_data.expiration_time = self._GetDateTimeValue(
+          record_values, 'Expires')
+      event_data.modification_time = self._GetDateTimeValue(
+          record_values, 'LastModified')
+      event_data.request_domain = record_values.get('RDomain', None)
+      parser_mediator.ProduceEventData(event_data)
+  
   def ParseContainersTable(
       self, parser_mediator, database=None, table=None, **unused_kwargs):
     """Parses a Containers table.
@@ -332,6 +439,10 @@ class MsieWebCacheESEDBPlugin(interface.ESEDBPlugin):
       esedb_table = database.GetTableByName(table_name)
       if esedb_table:
         self._ParseContainerTable(parser_mediator, esedb_table, container_name)
+      cookie_table_name = 'CookieEntryEx_{0:d}'.format(container_identifier)
+      cookie_table = database.GetTableByName(cookie_table_name)
+      if cookie_table.name==cookie_table_name:
+        self._ParseCookieExTable(parser_mediator, cookie_table)
 
   def ParseLeakFilesTable(
       self, parser_mediator, database=None, table=None, **unused_kwargs):
